refine clone_nore
whenever clone nore if there are no nore existing

+++ bootstrap.sh
@@ -94,11 +94,28 @@ NORE_CONFIGURE="${NORE_WORK%/}/configure"
 clone_nore() {
 	local n="`( cd ${PREFIX} && git remote -v 2>/dev/null | \
 						 		grep 'nore\.git' &>/dev/null; echo $? )`"
+	local t=0
+
 	if [ 0 -eq $n ]; then
-		`( cd ${PREFIX} && git reset --hard &>/dev/null )`
-		cd ${PREFIX} && git pull origin ${NORE_BRANCH} &>/dev/null
+		if [ "yes" = "$NORE_UPGRADE" ]; then
+			echo -n " + upgrading nore ... "
+			`( cd ${PREFIX} && git reset --hard &>/dev/null )`
+			cd ${PREFIX} && git pull origin ${NORE_BRANCH} &>/dev/null
+			t=$?
+		else
+			return 0
+		fi
 	else
-		git clone --depth=1 --branch=${NORE_BRANCH} ${GITHUB_H}/nore.git ${PREFIX} &>/dev/null
+		echo -n " + cloning nore ... "
+		git clone --depth=1 --branch=${NORE_BRANCH} \
+			${GITHUB_H}/nore.git ${PREFIX} &>/dev/null
+		t=$?
+	fi
+
+	if [ 0 -eq $t ]; then
+		echo "ok"
+	else
+		echo "failed"
 	fi
 }
 
@@ -172,24 +189,12 @@ END
 echo -n " + checking nore ... "
 if [ -x "$NORE_CONFIGURE" ]; then
 	echo "found"
-	if [ "yes" = "$NORE_UPGRADE" ]; then
-		echo -n " + upgrading nore ... "
-		if `clone_nore`; then
-			echo "ok"
-		else
-			echo "failed"
-		fi
-	fi
 else
 	echo "no found"
-	echo -n " + cloning nore ... "
-	if `clone_nore`; then
-		echo "ok"
-	else
-		echo "failed"
-	fi
 fi
 cat_configure
+clone_nore
+
 
 END="`date +%s`"
 echo
